LPS-109973 Remove SPICommentResource
