+++ modules/apps/headless/headless-common-spi/bnd.bnd
@@ -1,6 +1,6 @@
 Bundle-Name: Liferay Headless Common SPI
 Bundle-SymbolicName: com.liferay.headless.common.spi
-Bundle-Version: 6.1.2
+Bundle-Version: 7.0.0
 Export-Package:\
 	com.liferay.headless.common.spi.odata.entity,\
 	com.liferay.headless.common.spi.resource,\
